% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_taxa.R
\name{collapse_taxa}
\alias{collapse_taxa}
\title{Subset a phyloseq object to a set of taxa.}
\usage{
collapse_taxa(
  ps_obj,
  taxa_to_keep,
  discard_other = FALSE,
  merged_label = "Other"
)
}
\arguments{
\item{ps_obj}{Phyloseq object}

\item{taxa_to_keep}{taxon ids (taxids) to be kept. These taxon ids need to
be part of \code{taxa_names(ps_obj)}}

\item{discard_other}{if \code{TRUE}, any taxon not in \code{taxa_to_keep} is
discard. If \code{FALSE}, they are collapsed into a single taxon labeled
\code{other_label}}

\item{merged_label}{Label for the new merged taxon}
}
\value{
A phyloseq object
}
\description{
This function takes a phyloseq object and a list of taxon ids to be kept,
and discards or merges all other taxa.
}
\details{
This function is essentially a wrapper around \link[phyloseq]{prune_taxa} and
\link[phyloseq]{merge_taxa}.

This function, together with \link[fantaxtic]{top_taxa} replaces
\link[fantaxtic]{get_top_taxa}. Identical output can be obtained by setting
\code{FUN = sum} in \link[fantaxtic]{top_taxa}.
}
\examples{
data(GlobalPatterns)

# Top 10 most abundant ASVs over all samples, collapse other ASVs into 'Low abundance'
top <- top_taxa(GlobalPatterns, 10)
ps_collapsed <- collapse_taxa(GlobalPatterns,
                              taxa_to_keep = top$top_taxa$taxid,
                              merged_label = "Low abundance")

# Top 10 most abundant ASVs over all samples, discard other taxa
top <- top_taxa(GlobalPatterns, 10)
ps_collapsed <- collapse_taxa(GlobalPatterns, taxa_to_keep = top$taxid,
                              discard_other = TRUE)

# Keep genus Clostridium, collapse all others into Other genera
ps_tmp <- subset_taxa(GlobalPatterns, Genus == "Clostridium")
taxids <- taxa_names(ps_tmp)
ps_collapsed <- collapse_taxa(GlobalPatterns, taxa_to_keep = taxids,
                              merged_label = "Other genera")

}
