% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_colors.R
\name{gen_colors}
\alias{gen_colors}
\title{Generate a color palette.}
\usage{
gen_colors(n, clr = "#6495ed")
}
\arguments{
\item{n}{The number of colors to generate.}

\item{clr}{The base color from which to generate other colors.}
}
\value{
A vector of \code{n} colors.
}
\description{
This function generates \code{n} maximally separated colors with equal
saturation and brightness from a base color.
}
\examples{
gen_colors(5)
gen_colors(5, "blue")
}
