% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_palette.R
\name{gen_palette}
\alias{gen_palette}
\title{Generate a color palette.}
\usage{
gen_palette(clr_tbl, clr_pal = NULL, base_clr = "#6495ed")
}
\arguments{
\item{clr_tbl}{A \code{data.frame} or \code{matrix} with two columns: the
first column gives the labels to which colors need to be assigned; the
second column gives the number of tints and shades that need to be generated
for that color.}

\item{clr_pal}{A color palette with with \eqn{n} colors, where
\eqn{n =} \code{nrow(clr_tbl)}.}

\item{base_clr}{The base color from which to generate a color palette.}
}
\value{
A list of \eqn{n} items, each containing \eqn{i_n} shades.
}
\description{
This function generates \eqn{i_1, i_2, ... i_n} shades and tints specified
in \code{clr.tbl} for \eqn{n} colors \code{clr.pal}. If no \code{clr.tbl}
is specified, this function generates \eqn{n} maximally separated colors
with equal saturation and brightness from a base color.
}
\examples{
clr_tbl <- data.frame(Var = c("Phylum A", "Phylum B"),
                      Freq = c(5, 8))
gen_palette(clr_tbl)
gen_palette(clr_tbl, clr_pal = c("red", "blue"))
gen_palette(clr_tbl, base_clr = "blue")
}
