% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gen_shades_tints-deprecated}
\alias{gen_shades_tints-deprecated}
\title{Generate shade and tint variants for a base color.}
\usage{
gen_shades_tints(i, clr = "#6495ed")
}
\arguments{
\item{i}{The number of tints and/or shades to be generated.}

\item{clr}{The base color to use (hex or named color).}

\item{incl.base}{Should the base color be part of the generated palette?}
}
\value{
\code{i} variants of a base color.
}
\description{
\code{gen_shades_tints} returns \code{i/2} shades and \code{i/2} tints
of a base color.
}
\examples{
gen_shades(3)
gen_tints(3)
gen_shades_tints(3)
gen_shades(3, "blue")
gen_tints(3, "blue")
gen_shades_tints(3, "blue")
}
\seealso{
\code{\link{fantaxtic-deprecated}}
}
