% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_unique_labels.R
\name{gen_uniq_lbls}
\alias{gen_uniq_lbls}
\title{Generate unique labels.}
\usage{
gen_uniq_lbls(lbls, sep_char = " ")
}
\arguments{
\item{lbls}{A vector of labels for which to generate unique names.}

\item{sep_char}{A (combination of) characters with which to separate labels
and their unique counts.}
}
\value{
A factor of unique labels.
}
\description{
This function generates unique labels for an input vector by adding a
count to labels of which multiple occurrences are found.
}
\examples{
lbls <- as.character(c("A", "B", "B", "C", "C", "C", "D"))
gen_uniq_lbls(lbls)
gen_uniq_lbls(lbls, sep_char = "_")

#Also works with factors or numerics
gen_uniq_lvls(as.factor(lbls))
lbls <- as.numeric(c(1, 2, 2, 3, 3, 3, 4))
gen_uniq_lbls(lbls)
}
