% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_top_taxa-deprecated}
\alias{get_top_taxa-deprecated}
\title{Get the most abundant taxa from a phyloseq object}
\usage{
get_top_taxa(physeq_obj, n, relative = T, discard_other, other_label)
}
\arguments{
\item{physeq_obj}{A phyloseq object with an \code{otu_table} and a
\code{tax_table}.}

\item{n}{The number of top taxa to subset to.}

\item{relative}{Select taxa based on relative abundance?
(default = \code{TRUE})}

\item{discard_other}{Discard remainig taxa? (default = \code{FALSE})}

\item{other_label}{Label to give to collapsed taxa.}
}
\value{
A phyloseq object
}
\description{
This function subsets a phyloseq object to the top \eqn{n} most abundant
taxa. Users can choose whether to select by absolute counts or relative
counts, and whether to discard or collapse remaining taxa.
}
\examples{
data(GlobalPatterns)
get_top_taxa(GlobalPatterns, 10)
get_top_taxa(GlobalPatterns, 10, relative = FALSE)
get_top_taxa(GlobalPatterns, 10, relative = TRUE, discard_other = TRUE)
get_top_taxa(GlobalPatterns, 10, relative = TRUE, other_label = "Non-abundant taxa")
}
\seealso{
\code{\link{fantaxtic-deprecated}}
}
