% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_label.R
\name{move_label}
\alias{move_label}
\title{Relevel a factor in a psmelt dataframe}
\usage{
move_label(psdf, col_name, label = "Other", pos = 0)
}
\arguments{
\item{psdf}{A psmelt dataframe}

\item{col_name}{The name of the factor column}

\item{label}{The factor level to be moved}

\item{pos}{The position to move the level to.}
}
\value{
A data.frame
}
\description{
Moves a factor level in a psmelt dataframe to the desired position.
}
\examples{
data(GlobalPatterns)

# Move "Other" to position 0
top <- top_taxa(GlobalPatterns)
psdf <- psmelt(top$ps_obj)
psdf <- move_label(psdf, col_name = "Phylum", label =  "Other", pos = 0)
levels(psdf$Phylum)
}
