% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_nested_labels.R
\name{move_nested_labels}
\alias{move_nested_labels}
\title{Relevel a nested factor in a psmelt dataframe}
\usage{
move_nested_labels(
  psdf,
  top_level,
  nested_level,
  top_merged_label = "Other",
  nested_label = "Other",
  pos = Inf
)
}
\arguments{
\item{psdf}{A psmelt dataframe}

\item{top_level}{The name of the top level factor column}

\item{nested_level}{The name of the nested level factor column}

\item{top_merged_label}{The label for the top level merged taxon}

\item{nested_label}{The nested label to be moved}

\item{pos}{The position to move the level to.}
}
\value{
A data.frame
}
\description{
Moves a nested factor level in a fantaxtic psmelt dataframe to the desired position.
}
\examples{
data(GlobalPatterns)

# Move "Other" to position Inf
top <- nested_top_taxa(GlobalPatterns, top_tax_level = "Phylum", nested_tax_level = "Species", n_top_taxa = 3, n_nested_taxa = 3)
psdf <- psmelt(top$ps_obj)
psdf <- move_label(psdf, col_name = "Phylum", label =  "Other", pos = 0)
psdf <- move_nested_labels(psdf, top_level = "Phylum", nested_level = "Species", pos = Inf)
levels(psdf$Species)
}
