% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_na_taxa.R
\name{name_na_taxa}
\alias{name_na_taxa}
\title{Name NA taxa}
\usage{
name_na_taxa(ps_obj, include_rank = T, na_label = "Unknown <tax> (<rank>)")
}
\arguments{
\item{ps_obj}{A phyloseq object}

\item{include_rank}{Whether to include the rank of the lowest known annotation}

\item{na_label}{The label to assign to each filled in NA annotation.}
}
\value{
A phyloseq object
}
\description{
This function fills in any NA annotation with the lowest known taxonomic annotation (\code{<tax>})
from a higher rank (\code{<rank>}).
}
\examples{
data(GlobalPatterns)

# Fill in names for NA taxa, including their rank
ps_tmp <- name_na_taxa(GlobalPatterns)
View(tax_table(ps_tmp))

# Leave the rank out and alter the label
ps_tmp <- name_na_taxa(GlobalPatterns, include_rank = F, na_label = "Unknown <tax>")
View(tax_table(ps_tmp))
}
