% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_taxa.R
\name{name_taxa}
\alias{name_taxa}
\title{Add names for missing taxon names in phyloseq objects}
\usage{
name_taxa(physeq_obj, label = "Unknown", other_label = NULL,
  species = FALSE, unique_rank = NULL, unique_sep = " ")
}
\arguments{
\item{physeq_obj}{A phyloseq object with a \code{tax_table}.}

\item{other_label}{The label(s) of samples whose names should not be altered.}

\item{species}{Generate a 'Genus species' (i.e. 'Escherichia Coli') label
for the species level?}

\item{unique_rank}{The taxonomic rank by which to generate unique labels
(added number) if desired (default = \code{"Unknown"}).}

\item{unique_sep}{The text character(s) by which to separate the annotation
and the unique number (only when \code{!is.null(unique_rank)}).}

\item{unknown_label}{Label to prepend the taxon name with (default =
\code{"Unknown"}).}
}
\description{
This function adds names for OTUs/ASVs with incomplete taxonomic
annotations, i.e. annotation is only available up to a certain
taxonomic rank. It replaces \code{NA} values with the lowest
available taxonomic annotation for an OTU/ASV and a label indicating
that the annotation is unknown. Species can be renamed to genus +
species if desired.
}
\examples{
data(GlobalPatterns)
name_na_taxa(GlobalPatterns)
name_na_taxa(GlobalPatterns, unknown_label = "Unannotated")
}
