% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_colors.R
\name{shuffle_colors}
\alias{shuffle_colors}
\title{Reshuffle a list of colors.}
\usage{
shuffle_colors(clrs)
}
\arguments{
\item{clrs}{A vector of colors to reshuffle.}
}
\value{
A reshuffled vector of \code{clrs}.
}
\description{
This function reshuffles a vector of colours so that in a list of \code{n}
colors, every color \code{i} in \code{\{1, n / 2\}} will be placed next
to color \code{i + n / 2}. This is useful for visualization of
automatically generated color palettes based upon for example gradients
or HSL color circles.
}
\examples{
shuffle_colors(gen_colors(6))
}
