% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_duplicate_taxa.R
\name{label_duplicate_taxa}
\alias{label_duplicate_taxa}
\title{Add unique labels to identical taxa}
\usage{
label_duplicate_taxa(
  ps_obj,
  tax_level = NULL,
  asv_as_id = T,
  duplicate_label = "<tax> <id>"
)
}
\arguments{
\item{ps_obj}{Phyloseq object}

\item{tax_level}{The taxonomic level at which to identify duplicates. Can be
set to \code{NULL}. See details.}

\item{asv_as_id}{use ASVs as ids, rather than counts?}

\item{duplicate_label}{String that specifies how to construct the label. Must
contain at least \code{<tax>} and \code{<id>}}
}
\value{
A phyloseq object
}
\description{
This function takes a phyloseq object and adds a label (\code{<id>}) to ASVs with the same
taxonomic annotation (\code{<tax>}). The label is either the ASV name (taken from row.names)
or the count of each duplicated taxon.
}
\details{
By default, this function finds the most specific non-NA annotation for each
ASV and adds labels to that level. However, it is also possible to add the
labels at a specific taxonomic level by specifying \code{tax_level = <rank>}.
}
\examples{
data(GlobalPatterns)

# Label the lowest non-NA level
ps_tmp <- label_duplicate_taxa(GlobalPatterns)
View(tax_table(ps_tmp))

# Use ASVs as ids rather than counts
ps_tmp <- label_duplicate_taxa(GlobalPatterns, asv_as_id = T, duplicate_label = "<tax> ASV <id>")
View(tax_table(ps_tmp))

# Label at species leve lonly
ps_tmp <- label_duplicate_taxa(GlobalPatterns, tax_level = "Species")
View(tax_table(ps_tmp))
}
