% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_taxa.R
\name{nested_top_taxa}
\alias{nested_top_taxa}
\title{Get the most abundant taxa over two nested levels}
\usage{
nested_top_taxa(
  ps_obj,
  top_tax_level,
  nested_tax_level,
  n_top_taxa = 1,
  n_nested_taxa = 1,
  merged_label = "Other",
  by_proportion = T,
  ...
)
}
\arguments{
\item{ps_obj}{A phyloseq object with an \code{otu_table} and a
\code{tax_table}.}

\item{top_tax_level}{The name of the top taxonomic rank in the phyloseq object}

\item{nested_tax_level}{The name of the nested taxonomic rank in the phyloseq object}

\item{n_top_taxa}{The number of top taxa to identify at the top level.}

\item{n_nested_taxa}{The number of top taxa to identify at the nested level.}

\item{by_proportion}{Converts absolute abundances to proportions before
calculating the summary statistic (default = \code{TRUE}).}

\item{...}{Additional arguments to be passed \code{top_taxa}
(e.g. \code{grouping = <string>, FUN = mean, na.rm = TRUE},).}
}
\value{
A tibble with the rank, taxon id, grouping factors, abundance summary
statistic and taxonomy.
}
\description{
This function identifies the top \eqn{n} taxa and the top \eqn{m} taxa at a
nested level in a phyloseq object. Users specify the
summary statistic that is used to rank the taxa, e.g. \code{sum}, \code{mean} or
\code{median}. Furthermore, it is possible to add one or more grouping
factors from the \code{tax_table} to get group-specific top \eqn{n,m} taxa.
The top taxa can be identified based on the absolute abundances or proportions.
}
\details{
This function, together with \code{\link[fantaxtic]{collapse_taxa}}, replaces
\code{\link[fantaxtic]{get_top_taxa}}.
}
\examples{
data(GlobalPatterns)

# Top 10 most abundant ASVs over all samples
top_taxa(GlobalPatterns, 10)

# Top 10 most abundant ASVs over all samples by median abundance
top_taxa(GlobalPatterns, 10, FUN = median, na.rm = T)

# Top 10 most abundant ASVs over all samples using absolute abundances
top_taxa(GlobalPatterns, 10, by_proportion = FALSE)

# Top 2 most abundant ASVs per sample
top_taxa(GlobalPatterns, 2, grouping = "sample_id")

# Top 2 most abundant ASVs per sample type
top_taxa(GlobalPatterns, 2, grouping = "SampleType")

# Top 2 most abundant ASVs per sample type and group
set.seed(1)
sample_data(GlobalPatterns)$group <- as.factor(rbinom(nsamples(GlobalPatterns), 1, .5))
top_taxa(GlobalPatterns, 2, grouping = c("SampleType", "group"))
}
